Algorithms/findLargestValueInEachTreeRow/src/Solution.java
import tree.TreeNode;

import java.util.ArrayList;
import java.util.LinkedList;
import java.util.List;

public class Solution {
    /**
     * 515. Find Largest Value in Each Tree Row
     * 解题思路：
     *  直接层序遍历。
     * @param root
     * @return
     */
    public List<Integer> largestValues(TreeNode root) {
        List<Integer> ans = new ArrayList<>();
        List<TreeNode> queue = new LinkedList<>();
        if (root != null) queue.add(root);
        while (!queue.isEmpty()) {
            int count = queue.size();
            int max = Integer.MIN_VALUE;
            for (int i = 0; i < count; i++) {
                TreeNode treeNode = queue.remove(0);
                if (max < treeNode.val) max = treeNode.val;
                if (treeNode.left != null) queue.add(treeNode.left);
                if (treeNode.right != null) queue.add(treeNode.right);
            }
            ans.add(max);
        }
        return ans;
    }
}
